Making the upload folder feature better
Earlier, the feature worked by zipping the folder with JSZip at the client, uploading the zip and then unzipping it server-side. Now, all the files are uploaded and then placed in their relative paths accordingly server-side.
This should make uploads much faster on low-power/weak machines and quite frankly, the old method was quite stupid.
This is not finished as of yet which is why it is in the uploads-test files.